added routers. Split functions into router files
