+++ main.py
@@ -1,9 +1,10 @@
-from enum import Enum
-from typing import Optional
+from fastapi import FastAPI
 
-from fastapi import FastAPI, status, Response
+from router import blog_get, blog_post
 
 app = FastAPI()
+app.include_router(blog_get.router)
+app.include_router(blog_post.router)
 
 
 @app.get('/')
@@ -11,51 +12,4 @@ def index():
     return {'message': 'Hello World'}
 
 
-# @app.get('/blog/all')
-# def get_all_blogs():
-#     return {'message': 'All blogs provided'}
 
-
-@app.get('/blog/all',
-         tags=['blog'],
-         summary='Retrieve all blogs',
-         description='This api call simulates fetching all blogs',
-         response_description='The list of available blogs'
-         )
-def get_all_blogs(page=1, page_size: Optional[int] = None):
-    return {'message': f'All {page_size} blogs on page {page}'}
-
-
-@app.get('/blog/{id}/comments/{comment_id}', tags=['blog', 'comment'])
-def get_comment(id: int, comment_id: int, valid: bool = True, username: Optional[str] = None):
-    """
-    Simulates retrieving a comment of a blog
-
-    - **id**: mandatory path parameter
-    - **comment_id**: mandatory path parameter
-    - **valid**: optional query parameter
-    - **username**: optional query parameter
-
-    """
-    return {'message': f'{id=} {comment_id=} {valid=} {username=}'}
-
-
-class BlogType(str, Enum):
-    short = 'short'
-    story = 'story'
-    howto = 'howto'
-
-
-@app.get('/blog/type/{blog_type}', tags=['blog'])
-def get_blog_type(blog_type: BlogType):
-    return {'message': f'Blog type {blog_type}'}
-
-
-@app.get('/blog/{id}', status_code=status.HTTP_200_OK, tags=['blog'])
-def get_blog(id: int, response: Response):
-    if id > 5:
-        response.status_code = status.HTTP_404_NOT_FOUND
-        return {'error': f'Blog {id} not found'}
-    else:
-        response.status_code = status.HTTP_200_OK
-        return {'message': f'Blog with id {id}'}
