+++ router/blog_post.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+
+router = APIRouter(
+    prefix='/blog',
+    tags=['blog'],
+)
+
+@router.post('/new')
+def create_blog():
+    pass
